front: move consts related to Scenario in views/Scenario
