+++ front/src/applications/operationalStudies/views/Scenario/consts.ts
